+++ packages/editor/editor-core/src/plugins/index.ts
@@ -50,3 +50,4 @@ export { default as breakoutPlugin } from './breakout';
 export { default as alignment } from './alignment';
 export { default as editorDisabledPlugin } from './editor-disabled';
 export { default as indentationPlugin } from './indentation';
+export { default as annotationPlugin } from './annotation';