+++ packages/elements/share/src/__tests__/unit/components/ShareFormSpec.tsx
@@ -3,12 +3,12 @@ import Form, { FormFooter, FormSection } from '@atlaskit/form';
 import { shallow } from 'enzyme';
 import * as React from 'react';
 import { FormattedMessage } from 'react-intl';
-import { CommentField } from '../../components/CommentField';
-import { ShareForm } from '../../components/ShareForm';
-import { ShareHeader } from '../../components/ShareHeader';
-import { UserPickerField } from '../../components/UserPickerField';
-import { CopyLinkButton } from '../../components/CopyLinkButton';
-import { messages } from '../../i18n';
+import { CommentField } from '../../../components/CommentField';
+import { ShareForm } from '../../../components/ShareForm';
+import { ShareHeader } from '../../../components/ShareHeader';
+import { UserPickerField } from '../../../components/UserPickerField';
+import { CopyLinkButton } from '../../../components/CopyLinkButton';
+import { messages } from '../../../i18n';
 import { renderProp } from '../_testUtils';
 
 describe('ShareForm', () => {