+++ packages/navigation/centralised-app-switcher/src/components/jira-app-switcher.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
 import { AppSwitcherWrapper, Item, Section, ManageButton } from '../primitives';
-import {
-  RecentContainersProvider,
-  CustomLinksProvider,
-} from '../providers/jira-data-providers';
+import { CustomLinksProvider } from '../providers/jira-data-providers';
+import { RecentContainersProvider } from '../providers/instance-data-providers';
 
 export default ({ cloudId }) => {
   return (
@@ -16,12 +14,12 @@ export default ({ cloudId }) => {
           {({ isLoading: isLoadingCustomLinks, data: customLinksData }) => (
             <AppSwitcherWrapper>
               {isLoadingRecentContainers ? (
-                'Loading First Section...'
+                'Loading Recent Containers...'
               ) : (
-                <Section title="First Section">
-                  <Item>First Item</Item>
-                  <Item>Second Item</Item>
-                  <Item>{JSON.stringify(recentContainersData)}</Item>
+                <Section title="Recent Containers">
+                  {recentContainersData.data.map(({ objectId, name }) => (
+                    <Item key={objectId}>{name}</Item>
+                  ))}
                 </Section>
               )}
               {isLoadingCustomLinks ? (